Improved handling of no underscore directories situation
When "make html/dirhtml" is run without there being any changes to the
files, then Sphinx-to-github is still run and used to throw an exception
indicating that there was no work to do.

This was a little extreme and has been reduced to an info message, when
verbose is set, and the code continues without trouble otherwise.